refactor(autoware_traffic_light_classifier): declare classifier params in the node
Move ROS parameter declaration for the three classifiers out of their adapter
implementations into a new node-side classifier_params translation unit (under src/,
alongside the node files -- declaring parameters is a Node concern, so it sits at the
node layer rather than in classifier/). Each adapter constructor now takes its plain
Config, built by the node via declare_hsv/cnn/lamp_config, so the classifier files no
longer call declare_parameter or read the label file.

First step of the by-concern de-Node refactor; no behavior change. colcon build and
colcon test pass (83 tests, 0 failures) on an RTX 5080 with the real TensorRT engines.
