+++ perception/autoware_traffic_light_classifier/src/classifier/cnn_classifier.hpp
@@ -96,7 +96,7 @@ private:
 class CNNClassifier : public ClassifierInterface
 {
 public:
-  explicit CNNClassifier(rclcpp::Node * node_ptr);
+  CNNClassifier(rclcpp::Node * node_ptr, const CNNConfig & config);
   virtual ~CNNClassifier() = default;
 
   bool getTrafficSignals(
