+++ perception/autoware_traffic_light_classifier/src/classifier/cnn_lamp_recognizer.hpp
@@ -200,7 +200,7 @@ private:
 class CnnLampRecognizer : public ClassifierInterface
 {
 public:
-  explicit CnnLampRecognizer(rclcpp::Node * node_ptr);
+  CnnLampRecognizer(rclcpp::Node * node_ptr, const CnnLampRecognizerConfig & config);
   ~CnnLampRecognizer() override = default;
 
   bool getTrafficSignals(
