+++ perception/autoware_traffic_light_classifier/test/test_cnn_classifier_adapter.cpp
@@ -30,6 +30,7 @@
 //
 
 #include "../src/classifier/cnn_classifier.hpp"
+#include "../src/classifier_params.hpp"
 
 #include <autoware/cuda_utils/cuda_gtest_utils.hpp>
 #include <opencv2/core.hpp>
@@ -116,7 +117,8 @@ protected:
     // "environment unavailable" -> skip (not fail), with a reason the test reports.
     try {
       node_ = std::make_shared<rclcpp::Node>("cnn_classifier_adapter_test", options);
-      classifier_ = std::make_shared<tl::CNNClassifier>(node_.get());
+      classifier_ =
+        std::make_shared<tl::CNNClassifier>(node_.get(), tl::declare_cnn_config(node_.get()));
     } catch (const std::exception & e) {
       skip_reason_ = std::string("CNNClassifier environment unavailable: ") + e.what();
       classifier_.reset();
