+++ perception/autoware_traffic_light_classifier/test/test_cnn_lamp_recognizer_adapter.cpp
@@ -36,6 +36,7 @@
 //
 
 #include "../src/classifier/cnn_lamp_recognizer.hpp"
+#include "../src/classifier_params.hpp"
 
 #include <ament_index_cpp/get_package_share_directory.hpp>
 #include <autoware/cuda_utils/cuda_gtest_utils.hpp>
@@ -166,7 +167,8 @@ protected:
     // GTEST_SKIP.
     try {
       node_ = std::make_shared<rclcpp::Node>("cnn_lamp_recognizer_adapter_test", options);
-      recognizer_ = std::make_shared<tl::CnnLampRecognizer>(node_.get());
+      recognizer_ =
+        std::make_shared<tl::CnnLampRecognizer>(node_.get(), tl::declare_lamp_config(node_.get()));
     } catch (const std::exception & e) {
       skip_reason_ = std::string("CnnLampRecognizer environment unavailable: ") + e.what();
       recognizer_.reset();
